Fixes abs method in GameStats, moves helpers to own modules and StatTracker
